+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Fixed bug in `Router.Phoenix.Router` where resource routes were not filtered correctly according to the path bindings
 * Added `Pow.Extension.Config.extension_modules/2`
 * Deprecated `Pow.Extension.Config.discover_modules/2`
+* Added support for custom metadata in `PowPersistentSession.Plug.Cookie`:
+  * `PowPersistentSession.Plug.Cookie.create/3` now stores a metadata keyword list that can be populated
+  * `PowPersistentSession.Plug.Cookie.create/3` will now, instead of adding `:session_fingerprint` to the metadata, populate the `:session_metadata` keyword list with `:fingerprint`
+  * `PowPersistentSession.Plug.Cookie.authenticate/2` will now populate session metadata with what exists in `:session_metadata` key for the persistent session metadata
+* `PowPersistentSession.Plug.Cookie.create/3` now ensures to delete the previous persistent session first, if one is found in cookies
 
 ## v1.0.14 (2019-10-29)